requirements.txt
asgiref==3.3.1
Django==3.0.12
django-leaflet==0.27.1
freeze==3.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
